feat(monaco): add monaco package
Initial setup of shared monaco package. 

Multiple places on our site (LE, Codebytes, Sandbox) require a code editor. The goal of this shared package is to make a code editor with a standardized configuration available in multiple applications. Read more in this [RFC](https://www.notion.so/codecademy/Monaco-editor-as-a-shared-package-1f4484db165b4abc8394c3556451ef6a)