fix: fixes a monaco editor display bug for lines below 14-17
There was some css added to monaco-editor that allowed cypress tests to pass without axe issues.  But this caused a bug with lines below the set height to not be shown in the editor.   Here, I've removed that css, and I've made a PR to monolith that skips that axe check until we are able to add it back in DISC-529.  Root cause: https://github.com/dequelabs/axe-core/issues/2806